Rename StorageTest to SqliteStorageTest
Right now that's more accurate. In the future we might want to have StorageTest
as a base class, but that'll need some adjustments to the build, since we're
just linking the _test.cpp file for each test, and since the Qt MOC stuff
doesn't work if we implement the base class for the test in a header.