Add clang-tidy for linting
Using the rather generous default configuration for now.

Unfortunately, this'll only work out of the box if Clang is also used for
compilation :/

The problem was that standard C++ includes, which clang-tidy appears to need,
were omitted when using g++. There's probably a way to get this working, but for
now it seems using Clang for most development is acceptable.