Rename Client to Daemon
The naming was getting awfully confusing, and now that we want to introduce a
"client" for the server's REST interface, it'd just be too ambiguous.

Since the client consists of a daemon and a UI, and this is the daemon's main
class, we figured that's the name to go for.