Fix testProcessSignupsHandlesDelegateCase
Well, that's a bit of an understatement. In fact, I've rewritten a good chunk of
our daemon/networking logic to be synchronous to achieve this. The logic itself
was _almost_ sound already, just a null check missing.

As for the rewrite, the issue with this test (and many before it) was actually
in the application logic: We weren't chaining futures correctly. As a result,
the test wouldn't wait for all processing to finish, and therefore the
assertions would fail.

One approach to fix this could have been to use our weird home-grown functions
like chain() to ensure every last future is awaited. I tried this, and the code
would have become pretty difficult to follow.

Another solution could have been to upgrade to Qt 6.4 to get better
future/promise primitives, but this might have made it difficult for us to stay
compatible with KDE - though that's just an assumption.

A final option we considered was to rewrite the whole daemon/networking logic to
more a more traditional signals/slots approach (possibly using QStateMachine),
but with all the consecutive network calls we make, this would have been a major
operation and possibly wouldn't have left the code in a state where it's at
least as easy (ish) to understand as before.

Discounting all of these for now, we went down the route of making the daemon
and networking logic synchronous. In this commit, I've only made the parts
synchronous that would help this particular test to pass. But in future, we can
get rid of Network::getRequest and Network::postRequest, rename the
corresponding sync functions, and change the rest of the logic to not return/use
any futures.